Work around aarch64 kernel quirk
If the syscallno is -1 (and -1 only) then arg1 gets trashed early
and set to -ENOSYS. This does prevent us from knowning what the
real arg1 value was, but luckily, we don't actually ever need it.

+++ src/Registers.h
@@ -183,7 +183,7 @@ public:
   // Access the registers holding system-call numbers, results, and
   // parameters.
 
-  intptr_t syscallno() const { return RR_GET_REG(eax, rax, x[8]); }
+  intptr_t syscallno() const { return (int)RR_GET_REG(eax, rax, x[8]); }
   void set_syscallno(intptr_t syscallno) { RR_SET_REG(eax, rax, x[8], syscallno); }
 
   /**
